requirements.txt
pytest>=5.2.1
aiven-client>=2.2.4
